Add support for Darwin platform
Previously, the code in `keep_alive.go` was Linux-specific due to the
constants it used from the 'syscall' package. With this commit, OS-specific
implementations of certain functions have been placed into two files
(`keep_alive_darwin.go` and `keep_alive_linux.go`). This enables the Go
compiler to chose the correct implementation for the current platform at
build-time.

The new Darwin-specific implementation was taken from the
felixge/tcpkeepalive project on GitHub.
