+++ keep_alive_linux.go
@@ -1,0 +1,18 @@
+package delugeclient
+
+import (
+	"os"
+	"syscall"
+)
+
+func setIdle(fd int, secs int) error {
+	return os.NewSyscallError("setsockopt", syscall.SetsockoptInt(fd, syscall.IPPROTO_TCP, syscall.TCP_KEEPIDLE, secs))
+}
+
+func setCount(fd int, n int) error {
+	return os.NewSyscallError("setsockopt", syscall.SetsockoptInt(fd, syscall.IPPROTO_TCP, syscall.TCP_KEEPCNT, n))
+}
+
+func setInterval(fd int, secs int) error {
+	return os.NewSyscallError("setsockopt", syscall.SetsockoptInt(fd, syscall.IPPROTO_TCP, syscall.TCP_KEEPINTVL, secs))
+}
